maintain displayable identifier for deleted records in event
